Fold ap curl into ap proxy as a positional subcommand
Replaces the standalone `ap curl --connection <id> <curl args>` with a
positional form on `ap proxy`: `ap proxy --connection <id> curl <curl
args>`. All ap proxy flags come before the literal `curl`; everything
after is forwarded to real curl verbatim.

This is structurally cleaner than the prior split:

 - Cobra's SetInterspersed(false) handles the flag boundary natively,
   so we no longer need DisableFlagParsing and a hand-rolled "is the
   first arg --connection?" check.
 - Curl flags that clashed with ours (e.g. both have --config) are no
   longer a concern — they're past the boundary and never see cobra.
 - One command, two modes (listener / one-shot), discriminated by the
   `curl` literal in args[0].

Removes the now-unused config.NewResolverFromDefaults helper.
